이태연 퀴즈 수정 및 삭제 기능 구현

+++ src/main/java/com/lms/domain/mafia/view/MafiaOutputView.java
@@ -3,7 +3,7 @@ package com.lms.domain.mafia.view;
 public class MafiaOutputView {
 
 
-    public void printTodayMafia(int userId) {
+    public void printTodayMafia(Long userId) {
         System.out.println("오늘의 마피아가 선정되었습니다! 게시판을 확인하세요");
         System.out.println("오늘의 마피아입니다 : " + userId);
     }
